7/08/2024 (add persistent layer and relevant Dao of rule tree)

+++ BigMarket-domain/src/main/java/org/example/domain/strategy/service/rule/tree/ILogicTreeNode.java
@@ -4,7 +4,7 @@ import org.example.domain.strategy.service.rule.tree.factory.DefaultTreeFactory;
 
 public interface ILogicTreeNode {
 
-    DefaultTreeFactory.TreeActionEntity logic(String userId, Long strategyId, Integer awardId);
+    DefaultTreeFactory.TreeActionEntity logic(String userId, Long strategyId, Long awardId);
 
 
 }
