+++ BigMarket-domain/src/main/java/org/example/domain/strategy/service/rule/tree/impl/RuleStockLogicTreeNode.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 @Component("rule_stock")
 public class RuleStockLogicTreeNode implements ILogicTreeNode {
     @Override
-    public DefaultTreeFactory.TreeActionEntity logic(String userId, Long strategyId, Integer awardId) {
+    public DefaultTreeFactory.TreeActionEntity logic(String userId, Long strategyId, Long awardId) {
         return DefaultTreeFactory.TreeActionEntity.builder()
                 .ruleLogicCheckTypeVO(RuleLogicCheckTypeVO.TAKE_OVER)
                 .build();
